+++ src/components/Calculators/Electronics/Basic/StandardResistance/ESeriesRow.vue
@@ -2,12 +2,12 @@
 
   <tr>
     <td>{{ eSeries.name }}</td>
-    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestResistance')" width=120></calc-value></td>
-    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestResistanceError')" width=80></calc-value></td>
-    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrLowerResistance')" width=120></calc-value></td>
-    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrLowerResistanceError')" width=80></calc-value></td>
-    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrHigherResistance')" width=120></calc-value></td>
-    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrHigherResistanceError')" width=80></calc-value></td>
+    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestResistance')" :width=120></calc-value></td>
+    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestResistanceError')" :width=80></calc-value></td>
+    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrLowerResistance')" :width=120></calc-value></td>
+    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrLowerResistanceError')" :width=80></calc-value></td>
+    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrHigherResistance')" :width=120></calc-value></td>
+    <td><calc-value :calcVar="calc.getVar(this.eSeries.name + 'ClosestEqualOrHigherResistanceError')" :width=80></calc-value></td>
   </tr>
 
 </template>
@@ -217,9 +217,7 @@
       return {
       }
     },
-    mounted () {
-      console.log('ESeriesRow mounted.')
-    }
+    mounted () {}
   }
 
 </script>